Enum aws_sdk_acmpca::Error

#[non_exhaustive]
pub enum Error {

    CertificateMismatchException(CertificateMismatchException),
    ConcurrentModificationException(ConcurrentModificationException),
    InvalidArgsException(InvalidArgsException),
    InvalidArnException(InvalidArnException),
    InvalidNextTokenException(InvalidNextTokenException),
    InvalidPolicyException(InvalidPolicyException),
    InvalidRequestException(InvalidRequestException),
    InvalidStateException(InvalidStateException),
    InvalidTagException(InvalidTagException),
    LimitExceededException(LimitExceededException),
    LockoutPreventedException(LockoutPreventedException),
    MalformedCsrException(MalformedCsrException),
    MalformedCertificateException(MalformedCertificateException),
    PermissionAlreadyExistsException(PermissionAlreadyExistsException),
    RequestAlreadyProcessedException(RequestAlreadyProcessedException),
    RequestFailedException(RequestFailedException),
    RequestInProgressException(RequestInProgressException),
    ResourceNotFoundException(ResourceNotFoundException),
    TooManyTagsException(TooManyTagsException),
    Unhandled(Box<dyn Error + Send + Sync + 'static>),

}

All possible error types for this service.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

CertificateMismatchException(CertificateMismatchException)

The certificate authority certificate you are importing does not comply with conditions specified in the certificate that signed it.

Tuple Fields of CertificateMismatchException

0: CertificateMismatchException

ConcurrentModificationException(ConcurrentModificationException)

A previous update to your private CA is still ongoing.

Tuple Fields of ConcurrentModificationException

0: ConcurrentModificationException

InvalidArgsException(InvalidArgsException)

One or more of the specified arguments was not valid.

Tuple Fields of InvalidArgsException

0: InvalidArgsException

InvalidArnException(InvalidArnException)

The requested Amazon Resource Name (ARN) does not refer to an existing resource.

Tuple Fields of InvalidArnException

0: InvalidArnException

InvalidNextTokenException(InvalidNextTokenException)

The token specified in the NextToken argument is not valid. Use the token returned from your previous call to ListCertificateAuthorities.

Tuple Fields of InvalidNextTokenException

0: InvalidNextTokenException

InvalidPolicyException(InvalidPolicyException)

The resource policy is invalid or is missing a required statement. For general information about IAM policy and statement structure, see Overview of JSON Policies.

Tuple Fields of InvalidPolicyException

0: InvalidPolicyException

InvalidRequestException(InvalidRequestException)

The request action cannot be performed or is prohibited.

Tuple Fields of InvalidRequestException

0: InvalidRequestException

InvalidStateException(InvalidStateException)

The state of the private CA does not allow this action to occur.

Tuple Fields of InvalidStateException

0: InvalidStateException

InvalidTagException(InvalidTagException)

The tag associated with the CA is not valid. The invalid argument is contained in the message field.

Tuple Fields of InvalidTagException

0: InvalidTagException

LimitExceededException(LimitExceededException)

An ACM Private CA quota has been exceeded. See the exception message returned to determine the quota that was exceeded.

Tuple Fields of LimitExceededException

0: LimitExceededException

LockoutPreventedException(LockoutPreventedException)

The current action was prevented because it would lock the caller out from performing subsequent actions. Verify that the specified parameters would not result in the caller being denied access to the resource.

Tuple Fields of LockoutPreventedException

0: LockoutPreventedException

MalformedCsrException(MalformedCsrException)

The certificate signing request is invalid.

Tuple Fields of MalformedCsrException

0: MalformedCsrException

MalformedCertificateException(MalformedCertificateException)

One or more fields in the certificate are invalid.

Tuple Fields of MalformedCertificateException

0: MalformedCertificateException

PermissionAlreadyExistsException(PermissionAlreadyExistsException)

The designated permission has already been given to the user.

Tuple Fields of PermissionAlreadyExistsException

0: PermissionAlreadyExistsException

RequestAlreadyProcessedException(RequestAlreadyProcessedException)

Your request has already been completed.

Tuple Fields of RequestAlreadyProcessedException

0: RequestAlreadyProcessedException

RequestFailedException(RequestFailedException)

The request has failed for an unspecified reason.

Tuple Fields of RequestFailedException

0: RequestFailedException

RequestInProgressException(RequestInProgressException)

Your request is already in progress.

Tuple Fields of RequestInProgressException

0: RequestInProgressException

ResourceNotFoundException(ResourceNotFoundException)

A resource such as a private CA, S3 bucket, certificate, audit report, or policy cannot be found.

Tuple Fields of ResourceNotFoundException

0: ResourceNotFoundException

TooManyTagsException(TooManyTagsException)

You can associate up to 50 tags with a private CA. Exception information is contained in the exception message field.

Tuple Fields of TooManyTagsException

0: TooManyTagsException

Unhandled(Box<dyn Error + Send + Sync + 'static>)

An unhandled error occurred.

Tuple Fields of Unhandled

0: Box<dyn Error + Send + Sync + 'static>