Module aws_sdk_acmpca::error

Modules

certificate_mismatch_exception

See CertificateMismatchException

concurrent_modification_exception

See ConcurrentModificationException

invalid_args_exception

See InvalidArgsException

invalid_arn_exception

See InvalidArnException

invalid_next_token_exception

See InvalidNextTokenException

invalid_policy_exception

See InvalidPolicyException

invalid_request_exception

See InvalidRequestException

invalid_state_exception

See InvalidStateException

invalid_tag_exception

See InvalidTagException

limit_exceeded_exception

See LimitExceededException

lockout_prevented_exception

See LockoutPreventedException

malformed_certificate_exception

See MalformedCertificateException

malformed_csr_exception

See MalformedCsrException

permission_already_exists_exception

See PermissionAlreadyExistsException

request_already_processed_exception

See RequestAlreadyProcessedException

request_failed_exception

See RequestFailedException

request_in_progress_exception

See RequestInProgressException

resource_not_found_exception

See ResourceNotFoundException

too_many_tags_exception

See TooManyTagsException

Structs

CertificateMismatchException

The certificate authority certificate you are importing does not comply with

ConcurrentModificationException

A previous update to your private CA is still ongoing.

CreateCertificateAuthorityAuditReportError

CreateCertificateAuthorityError

CreatePermissionError

DeleteCertificateAuthorityError

DeletePermissionError

DeletePolicyError

DescribeCertificateAuthorityAuditReportError

DescribeCertificateAuthorityError

GetCertificateAuthorityCertificateError

GetCertificateAuthorityCsrError

GetCertificateError

GetPolicyError

ImportCertificateAuthorityCertificateError

InvalidArgsException

One or more of the specified arguments was not valid.

InvalidArnException

The requested Amazon Resource Name (ARN) does not refer to an existing

InvalidNextTokenException

The token specified in the NextToken argument is not valid. Use the token

InvalidPolicyException

The resource policy is invalid or is missing a required statement. For general

InvalidRequestException

The request action cannot be performed or is prohibited.

InvalidStateException

The state of the private CA does not allow this action to occur.

InvalidTagException

The tag associated with the CA is not valid. The invalid argument is contained in the

IssueCertificateError

LimitExceededException

An ACM Private CA quota has been exceeded. See the exception message returned to determine the

ListCertificateAuthoritiesError

ListPermissionsError

ListTagsError

LockoutPreventedException

The current action was prevented because it would lock the caller out from performing

MalformedCertificateException

One or more fields in the certificate are invalid.

MalformedCsrException

The certificate signing request is invalid.

PermissionAlreadyExistsException

The designated permission has already been given to the user.

PutPolicyError

RequestAlreadyProcessedException

Your request has already been completed.

RequestFailedException

The request has failed for an unspecified reason.

RequestInProgressException

Your request is already in progress.

ResourceNotFoundException

A resource such as a private CA, S3 bucket, certificate, audit report, or policy

RestoreCertificateAuthorityError

RevokeCertificateError

TagCertificateAuthorityError

TooManyTagsException

You can associate up to 50 tags with a private CA. Exception information is contained

UntagCertificateAuthorityError

UpdateCertificateAuthorityError

Enums

CreateCertificateAuthorityAuditReportErrorKind

CreateCertificateAuthorityErrorKind

CreatePermissionErrorKind

DeleteCertificateAuthorityErrorKind

DeletePermissionErrorKind

DeletePolicyErrorKind

DescribeCertificateAuthorityAuditReportErrorKind

DescribeCertificateAuthorityErrorKind

GetCertificateAuthorityCertificateErrorKind

GetCertificateAuthorityCsrErrorKind

GetCertificateErrorKind

GetPolicyErrorKind

ImportCertificateAuthorityCertificateErrorKind

IssueCertificateErrorKind

ListCertificateAuthoritiesErrorKind

ListPermissionsErrorKind

ListTagsErrorKind

PutPolicyErrorKind

RestoreCertificateAuthorityErrorKind

RevokeCertificateErrorKind

TagCertificateAuthorityErrorKind

UntagCertificateAuthorityErrorKind

UpdateCertificateAuthorityErrorKind