Module aws_sdk_acmpca::model

Modules

access_description

See AccessDescription

access_method

See AccessMethod

api_passthrough

See ApiPassthrough

asn1_subject

See Asn1Subject

certificate_authority

See CertificateAuthority

certificate_authority_configuration

See CertificateAuthorityConfiguration

crl_configuration

See CrlConfiguration

csr_extensions

See CsrExtensions

edi_party_name

See EdiPartyName

extended_key_usage

See ExtendedKeyUsage

extensions

See Extensions

general_name

See GeneralName

key_usage

See KeyUsage

ocsp_configuration

See OcspConfiguration

other_name

See OtherName

permission

See Permission

policy_information

See PolicyInformation

policy_qualifier_info

See PolicyQualifierInfo

qualifier

See Qualifier

revocation_configuration

See RevocationConfiguration

tag

See Tag

validity

See Validity

Structs

AccessDescription

Provides access information used by the authorityInfoAccess and

AccessMethod

Describes the type and format of extension access. Only one of

ApiPassthrough

Contains X.509 certificate information to be placed in an issued certificate. An

Asn1Subject

Contains information about the certificate subject. The Subject field in

CertificateAuthority

Contains information about your private certificate authority (CA). Your private CA

CertificateAuthorityConfiguration

Contains configuration information for your private certificate authority (CA). This

CrlConfiguration

Contains configuration information for a certificate revocation list (CRL). Your

CsrExtensions

Describes the certificate extensions to be added to the certificate signing request

EdiPartyName

Describes an Electronic Data Interchange (EDI) entity as described in as defined in

ExtendedKeyUsage

Specifies additional purposes for which the certified public key may be used other

Extensions

Contains X.509 extension information for a certificate.

GeneralName

Describes an ASN.1 X.400 GeneralName as defined in RFC 5280. Only one of the

KeyUsage

Defines one or more purposes for which the key contained in the certificate can be

OcspConfiguration

Contains information to enable and configure Online Certificate Status Protocol (OCSP) for

OtherName

Defines a custom ASN.1 X.400 GeneralName using an object identifier (OID)

Permission

Permissions designate which private CA actions can be performed by an AWS service or

PolicyInformation

Defines the X.509 CertificatePolicies extension.

PolicyQualifierInfo

Modifies the CertPolicyId of a PolicyInformation object with

Qualifier

Defines a PolicyInformation qualifier. ACM Private CA supports the certification practice

RevocationConfiguration

Certificate revocation information used by the CreateCertificateAuthority and UpdateCertificateAuthority actions. Your private certificate authority (CA)

Tag

Tags are labels that you can use to identify and organize your private CAs. Each tag

Validity

Validity specifies the period of time during which a certificate is valid. Validity

Enums

AccessMethodType

ActionType

AuditReportResponseFormat

AuditReportStatus

CertificateAuthorityStatus

CertificateAuthorityType

ExtendedKeyUsageType

FailureReason

KeyAlgorithm

KeyStorageSecurityStandard

PolicyQualifierId

ResourceOwner

NOTE: ResourceOwner::Self has been renamed to ::SelfValue.

RevocationReason

S3ObjectAcl

SigningAlgorithm

ValidityPeriodType