Module aws_sdk_acmpca::operation

Structs

CreateCertificateAuthority

Creates a root or subordinate private certificate authority (CA). You must specify the

CreateCertificateAuthorityAuditReport

Creates an audit report that lists every time that your CA private key is used. The

CreatePermission

Grants one or more permissions on a private CA to the AWS Certificate Manager (ACM) service

DeleteCertificateAuthority

Deletes a private certificate authority (CA). You must provide the Amazon Resource

DeletePermission

Revokes permissions on a private CA granted to the AWS Certificate Manager (ACM) service principal

DeletePolicy

Deletes the resource-based policy attached to a private CA. Deletion will remove any

DescribeCertificateAuthority

Lists information about your private certificate authority (CA) or one that has been

DescribeCertificateAuthorityAuditReport

Lists information about a specific audit report created by calling the CreateCertificateAuthorityAuditReport action. Audit information is created

GetCertificate

Retrieves a certificate from your private CA or one that has been shared with you. The

GetCertificateAuthorityCertificate

Retrieves the certificate and certificate chain for your private certificate authority

GetCertificateAuthorityCsr

Retrieves the certificate signing request (CSR) for your private certificate authority

GetPolicy

Retrieves the resource-based policy attached to a private CA. If either the private CA

ImportCertificateAuthorityCertificate

Imports a signed private CA certificate into ACM Private CA. This action is used when you are

IssueCertificate

Uses your private certificate authority (CA), or one that has been shared with you, to

ListCertificateAuthorities

Lists the private certificate authorities that you created by using the CreateCertificateAuthority action.

ListPermissions

List all permissions on a private CA, if any, granted to the AWS Certificate Manager (ACM) service

ListTags

Lists the tags, if any, that are associated with your private CA or one that has been

PutPolicy

Attaches a resource-based policy to a private CA.

RestoreCertificateAuthority

Restores a certificate authority (CA) that is in the DELETED state. You

RevokeCertificate

Revokes a certificate that was issued inside ACM Private CA. If you enable a certificate

TagCertificateAuthority

Adds one or more tags to your private CA. Tags are labels that you can use to identify

UntagCertificateAuthority

Remove one or more tags from your private CA. A tag consists of a key-value pair. If

UpdateCertificateAuthority

Updates the status or configuration of a private certificate authority (CA). Your